# Refactor zero-copy buffers for performance and to prevent memory leaks 