--- a/core/lib.rs
+++ b/core/lib.rs
@@ -16,6 +16,7 @@ pub use crate::isolate::*;
 pub use crate::js_errors::*;
 pub use crate::libdeno::deno_buf;
 pub use crate::libdeno::deno_mod;
+pub use crate::libdeno::PinnedBuf;
 pub use crate::modules::*;
 
 pub fn v8_version() -> &'static str {