@piscisaureus 2019-04-30
@ry I tucked another huge refactor to the end, so if you care give it another look.
There's a lot more deno_buf related code that can be removed now but I want to do that in a follow-up PR, the diff is pretty big already.