@ry 2019-05-01
this branch
```
~/src/deno> ./third_party/wrk/mac/wrk -d 20s http://127.0.0.1:4544/
Running 20s test @ http://127.0.0.1:4544/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   195.79us  358.70us  19.82ms   99.78%
    Req/Sec    25.46k     1.91k   27.83k    83.08%
  1018097 requests in 20.10s, 49.52MB read
Requests/sec:  50652.61
Transfer/sec:      2.46MB
```
master (c36b5d)
```
~/src/deno> ./third_party/wrk/mac/wrk -d 20s http://127.0.0.1:4544/
Running 20s test @ http://127.0.0.1:4544/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   205.82us   60.99us   4.50ms   93.11%
    Req/Sec    23.21k     1.84k   26.35k    72.89%
  928136 requests in 20.10s, 45.14MB read
Requests/sec:  46176.63
Transfer/sec:      2.25MB
```
node v12.1.0 using `tools/node_tcp.js`
```
~/src/deno> ./third_party/wrk/mac/wrk -d 20s http://127.0.0.1:4544/
Running 20s test @ http://127.0.0.1:4544/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   235.46us  155.12us   9.81ms   91.76%
    Req/Sec    20.78k     1.89k   27.41k    87.06%
  831459 requests in 20.10s, 40.44MB read
Requests/sec:  41365.67
Transfer/sec:      2.01MB
```